requirements.txt
pillow==9.0.0
flask==2.0.2
gtfs-realtime-bindings==0.0.7
requests==2.27.1
nyct_gtfs==1.2.0
pytz==2021.3
sentry_sdk==1.5.4
num2words==0.5.10
spotipy==2.19.0
python-dotenv==0.19.2
pyYAML==6.0
adafruit-io==2.6.0
pep8==1.7.1
pylint==2.13.5
APScheduler==3.9.1